+++ src/common/injectable-interfaces.ts
@@ -11,7 +11,7 @@ export interface InjectableResolver {
     resolver(): IResolvers;
 }
 
-export interface LogConfig {
+export interface LoggerConfig {
     getLogConfiguration(): LoggerConfiguration;
 }
 export interface HeaderConfig {
@@ -19,5 +19,5 @@ export interface HeaderConfig {
 }
 export interface PolarisServerConfig {
     getPolarisProperties(): PolarisProperties;
-    getApplicationLogProperties(): ApplicationLogProperties;
+    getPropertiesForLog(): ApplicationLogProperties;
 }