+++ src/http/request/polaris-request-headers.ts
@@ -1,17 +1,16 @@
 import { HeaderNames } from '../header-names';
 
 export class PolarisRequestHeaders {
-    readonly dataVersion: number;
-    readonly isSnapshot: boolean;
-    readonly includeLinkedOperation: boolean;
-    readonly snapshotPageSize: number;
-    readonly isPolling: boolean;
-    readonly requestId: string;
-    readonly upn: string;
-    readonly eventKind: string;
-    readonly realityId: number;
-    readonly requestingSystemId: string;
-    readonly requestingSystemName: string;
+    readonly dataVersion?: number;
+    readonly isSnapshot?: boolean;
+    readonly includeLinkedOperation?: boolean;
+    readonly snapshotPageSize?: number;
+    readonly requestId?: string;
+    readonly upn?: string;
+    readonly eventKind?: string;
+    readonly realityId?: string;
+    readonly requestingSystemId?: string;
+    readonly requestingSystemName?: string;
 
     constructor(headers: any) {
         const joi = require('joi');
@@ -22,11 +21,10 @@ export class PolarisRequestHeaders {
             joi.boolean(),
         );
         this.snapshotPageSize = joi.attempt(headers[HeaderNames.SNAPSHOT_PAGE_SIZE], joi.number());
-        this.isPolling = joi.attempt(headers[HeaderNames.POLLING], joi.boolean());
         this.requestId = joi.attempt(headers[HeaderNames.REQUEST_ID], joi.string());
         this.upn = joi.attempt(headers[HeaderNames.UPN], joi.string());
         this.eventKind = joi.attempt(headers[HeaderNames.EVENT_KIND], joi.string());
-        this.realityId = joi.attempt(headers[HeaderNames.REALITY_ID], joi.number());
+        this.realityId = joi.attempt(headers[HeaderNames.REALITY_ID], joi.string());
         this.requestingSystemId = joi.attempt(headers[HeaderNames.REQUESTING_SYS], joi.string());
         this.requestingSystemName = joi.attempt(
             headers[HeaderNames.REQUESTING_SYS_NAME],