+++ src/inversion-of-control/polaris-types.ts
@@ -1,9 +1,11 @@
 import {
+    HeaderConfig,
     InjectableResolver,
     InjectableType,
-    LogConfig,
+    LoggerConfig,
     PolarisServerConfig,
 } from '../common/injectable-interfaces';
+import { GraphqlLogger } from '../logging/graphql-logger';
 import { PolarisMiddleware } from '../middlewares/polaris-middleware';
 import { SchemaCreator } from '../schema/utils/schema.creator';
 import { GraphQLServer } from '../server/graphql.server';
@@ -11,11 +13,11 @@ import { GraphQLServer } from '../server/graphql.server';
 export const POLARIS_TYPES = {
     GraphQLServer: Symbol('GraphQLServer'),
     SchemaCreator: Symbol('SchemaCreator'),
-    PolarisLogger: Symbol('PolarisLogger'),
+    GraphqlLogger: Symbol('GraphqlLogger'),
     PolarisMiddleware: Symbol('PolarisMiddleware'),
     PolarisServerConfig: Symbol('PolarisServerConfig'),
     HeaderConfig: Symbol('HeaderConfig'),
-    LogConfig: Symbol('LogConfig'),
+    LoggerConfig: Symbol('LoggerConfig'),
     InjectableType: Symbol('InjectableType'),
     InjectableResolver: Symbol('InjectableResolver'),
     CommonEntityInterface: Symbol('CommonEntityInterface`'),