+++ src/middlewares/polaris-middleware-creator.ts
@@ -1,17 +1,20 @@
 import { GraphQLResolveInfo } from 'graphql';
-import { PolarisContext, PolarisMiddleware } from './polaris-middleware';
+import { PolarisContext } from '../server/polaris-context';
+import { PolarisMiddleware } from './polaris-middleware';
 
-export const createMiddleware = (middleware: PolarisMiddleware) => {
+export const createMiddleware = <TContext extends PolarisContext = PolarisContext>(
+    middleware: PolarisMiddleware<TContext>,
+) => {
     return async (
         resolve: (
             root: any,
             args: { [argName: string]: any },
-            context: PolarisContext,
+            context: TContext,
             info: GraphQLResolveInfo,
         ) => string,
         root: any,
         args: { [argName: string]: any },
-        context: PolarisContext,
+        context: TContext,
         info: GraphQLResolveInfo,
     ) => {
         middleware.preResolve({ root, args, context, info });