+++ src/server/polaris-context.ts
@@ -1,0 +1,6 @@
+import { PolarisRequestHeaders } from '../http/request/polaris-request-headers';
+
+export interface PolarisContext {
+    headers: PolarisRequestHeaders;
+    body: any;
+}