+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# [1.3.0](https://github.com/Enigmatis/polaris/compare/v1.2.1...v1.3.0) (2019-01-31)
+
+
+### Features
+
+* **log:** print headers to log ([3b304a4](https://github.com/Enigmatis/polaris/commit/3b304a4))
+* **log:** print headers to log ([a431bd9](https://github.com/Enigmatis/polaris/commit/a431bd9))
+
+## [1.2.1](https://github.com/Enigmatis/polaris/compare/v1.2.0...v1.2.1) (2019-01-31)
+
+
+### Bug Fixes
+
+* jest configuration standard fixes ([2fcf134](https://github.com/Enigmatis/polaris/commit/2fcf134))
+* pre-commit changes ([932b74b](https://github.com/Enigmatis/polaris/commit/932b74b))
+* removed tsx and transform in jest config ([c91fb78](https://github.com/Enigmatis/polaris/commit/c91fb78))
+
 # [1.2.0](https://github.com/Enigmatis/polaris/compare/v1.1.0...v1.2.0) (2019-01-30)